[RENAME] Rename PhotoGridVC to ChatVC, MoreVC to PhotoGridVC
