+++ IN_SOPT_Third_Task/IN_SOPT_Third_Task/Sources/ViewControllers/KakaoTabbarController.swift
@@ -54,7 +54,7 @@ extension KakaoTabbarController : UITabBarControllerDelegate {
         let friendNavigationController =  UINavigationController(rootViewController: friendViewController)
         friendNavigationController.navigationBar.isHidden = true
         
-        let chatViewController = PhotoGridViewController()
+        let chatViewController = ChatViewController()
         let chatNavigationController = UINavigationController(rootViewController: chatViewController)
         chatNavigationController.navigationBar.isHidden = true
